+++ components/clientComponents/forms/Form/Form.tsx
@@ -1,26 +1,25 @@
 "use client";
-import React, { useEffect, useState, useRef, type JSX, useCallback } from "react";
-import { FormikProps, withFormik } from "formik";
+import React, { useEffect, useState, useRef, useCallback } from "react";
+import { withFormik } from "formik";
 import { getFormInitialValues } from "@lib/formBuilder";
 import { getErrorList, setFocusOnErrorMessage, validateOnSubmit } from "@lib/validation/validation";
 import { Alert, RichText } from "@clientComponents/forms";
-import { Button } from "@clientComponents/globals";
+
+import { type FormProps, type InnerFormProps } from "./types";
+import { type Language } from "@lib/types/form-builder-types";
+import { type Responses } from "@lib/types";
+
 import { logMessage } from "@lib/logger";
 import { useTranslation } from "@i18n/client";
-import { TFunction } from "i18next";
+
 import Loader from "../../globals/Loader";
-import { cn } from "@lib/utils";
-import { Responses, PublicFormRecord, Validate } from "@lib/types";
+
 import { ErrorStatus } from "../Alert/Alert";
 import { submitForm } from "app/(gcforms)/[locale]/(form filler)/id/[...props]/actions";
-import useFormTimer from "@lib/hooks/useFormTimer";
 import { useFormValuesChanged } from "@lib/hooks/useValueChanged";
 import { useGCFormsContext } from "@lib/hooks/useGCFormContext";
 import { Review } from "../Review/Review";
 import { LockedSections } from "@formBuilder/components/shared/right-panel/treeview/types";
-import { BackButton } from "@formBuilder/[id]/preview/BackButton";
-import { Language } from "@lib/types/form-builder-types";
-import { BackButtonGroup } from "../BackButtonGroup/BackButtonGroup";
 import { StatusError } from "../StatusError/StatusError";
 import {
   removeFormContextValues,
@@ -30,115 +29,9 @@ import { filterShownElements, filterValuesByShownElements } from "@lib/formConte
 import { formHasGroups } from "@lib/utils/form-builder/formHasGroups";
 import { showReviewPage } from "@lib/utils/form-builder/showReviewPage";
 import { useFormDelay } from "@lib/hooks/useFormDelayContext";
-import { SaveAndResume } from "@clientComponents/forms/SaveAndResume/SaveAndResume";
 
-interface SubmitButtonProps {
-  getFormDelay: () => number;
-  formID: string;
-  formTitle: string;
-}
-const SubmitButton: React.FC<SubmitButtonProps> = ({ getFormDelay, formID, formTitle }) => {
-  const { t } = useTranslation();
-  const [formTimerState, { startTimer, checkTimer, disableTimer }] = useFormTimer();
-  const [submitTooEarly, setSubmitTooEarly] = useState(false);
-  const screenReaderRemainingTime = useRef(formTimerState.remainingTime);
-  const formDelay = useRef(getFormDelay());
-
-  // If the formDelay is less than 0 or the app is in test mode, disable the timer
-  // because the user has already spent enough time on the form.
-
-  const formTimerEnabled = process.env.NEXT_PUBLIC_APP_ENV !== "test" && formDelay.current > 0;
-
-  // The empty array of dependencies ensures that this useEffect only runs once on mount
-  useEffect(() => {
-    if (formTimerEnabled) {
-      logMessage.debug(`Starting Form Timer with delay: ${formDelay.current}`);
-      startTimer(formDelay.current);
-    } else {
-      disableTimer();
-    }
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, []);
-
-  useEffect(() => {
-    if (formTimerEnabled && formTimerState.remainingTime > 0) {
-      // Initiate a callback to ensure that state of submit button is correctly displayed
-
-      // Calling the checkTimer modifies the state of the formTimerState
-      // Which recalls this useEffect at least every second
-      const timerID = setTimeout(() => checkTimer(), 1000);
-
-      return () => {
-        clearTimeout(timerID);
-      };
-    }
-  }, [checkTimer, formTimerState.remainingTime, formTimerEnabled]);
-
-  return (
-    <>
-      <div
-        className={cn({
-          "border-l-2": submitTooEarly,
-          "border-red-default": submitTooEarly && formTimerState.remainingTime > 0,
-          "border-green-default": submitTooEarly && formTimerState.remainingTime === 0,
-          "pl-3": submitTooEarly,
-        })}
-      >
-        {submitTooEarly &&
-          (formTimerState.remainingTime > 0 ? (
-            <>
-              <div role="alert" className="gc-label text-red-default">
-                {t("spam-error.error-part-1")} {formTimerState.timerDelay}{" "}
-                {t("spam-error.error-part-2")}
-                <span className="sr-only">
-                  {" "}
-                  {t("spam-error.prompt-part-1")} {screenReaderRemainingTime.current}{" "}
-                  {t("spam-error.prompt-part-2")}
-                </span>
-              </div>
-              <div aria-hidden={true} className="gc-description">
-                {t("spam-error.prompt-part-1")} {formTimerState.remainingTime}{" "}
-                {t("spam-error.prompt-part-2")}
-              </div>
-            </>
-          ) : (
-            <div role="alert">
-              <p className="gc-label text-green-default">{t("spam-error.success-message")}</p>
-              <p className="gc-description">{t("spam-error.success-prompt")}</p>
-            </div>
-          ))}
-      </div>
-      <Button
-        id="form-submit-button"
-        type="submit"
-        onClick={(e) => {
-          if (formTimerEnabled) checkTimer();
-          screenReaderRemainingTime.current = formTimerState.remainingTime;
-          if (!formTimerState.canSubmit) {
-            e.preventDefault();
-            window.dataLayer = window.dataLayer || [];
-            window.dataLayer.push({
-              event: "form_submission_spam_trigger",
-              formID: formID,
-              formTitle: formTitle,
-              submitTime: formTimerState.remainingTime,
-            });
-
-            setSubmitTooEarly(true);
-            // In case the useEffect timer failed check again
-            return;
-          }
-          // Only change state if submitTooEarly is already set to true
-          submitTooEarly && setSubmitTooEarly(false);
-        }}
-      >
-        {t("submitButton")}
-      </Button>
-    </>
-  );
-};
-
-type InnerFormProps = FormProps & FormikProps<Responses>;
+import { FormActions } from "./FormActions";
+import { PrimaryFormButtons } from "./PrimaryFormButtons";
 
 /**
  * This is the "inner" form component that isn't connected to Formik and just renders a simple form
@@ -150,9 +43,11 @@ const InnerForm: React.FC<InnerFormProps> = (props) => {
     handleSubmit,
     status,
     language,
-    formRecord: { id: formID, form, securityAttribute },
+    formRecord: { id: formID, form },
     dirty,
   }: InnerFormProps = props;
+
+  const { t } = useTranslation();
   const [canFocusOnError, setCanFocusOnError] = useState(false);
   const [lastSubmitCount, setLastSubmitCount] = useState(-1);
 
@@ -161,13 +56,10 @@ const InnerForm: React.FC<InnerFormProps> = (props) => {
   const isShowReviewPage = showReviewPage(form);
   const showIntro = isGroupsCheck ? currentGroup === LockedSections.START : true;
   const groupsHeadingRef = useRef<HTMLHeadingElement>(null);
-
-  const { t } = useTranslation();
+  const { getFormDelayWithGroups, getFormDelayWithoutGroups } = useFormDelay();
 
   // Used to set any values we'd like added for use in the below withFormik handleSubmit().
   useFormValuesChanged();
-
-  const { getFormDelayWithGroups, getFormDelayWithoutGroups } = useFormDelay();
 
   const errorList = props.errors ? getErrorList(props) : null;
   const errorId = "gc-form-errors";
@@ -310,89 +202,34 @@ const InnerForm: React.FC<InnerFormProps> = (props) => {
               <Review language={language as Language} />
             )}
 
-            <div className="flex">
-              {isGroupsCheck && isShowReviewPage && (
-                <BackButtonGroup
-                  language={language as Language}
-                  onClick={() => groupsHeadingRef.current?.focus()}
-                />
-              )}
-              {props.renderSubmit ? (
-                props.renderSubmit({
-                  validateForm: props.validateForm,
-                  fallBack: () => {
-                    return (
-                      <div>
-                        {isGroupsCheck && isShowReviewPage && (
-                          <BackButton
-                            language={language as Language}
-                            onClick={() => groupsHeadingRef.current?.focus()}
-                          />
-                        )}
-                        <div className="inline-block">
-                          <SubmitButton
-                            getFormDelay={() =>
-                              isShowReviewPage
-                                ? getFormDelayWithGroups()
-                                : getFormDelayWithoutGroups(form.elements)
-                            }
-                            formID={formID}
-                            formTitle={form.titleEn}
-                          />
-                        </div>
-                      </div>
-                    );
-                  },
-                })
-              ) : (
-                <SubmitButton
-                  getFormDelay={() =>
-                    isShowReviewPage
-                      ? getFormDelayWithGroups()
-                      : getFormDelayWithoutGroups(form.elements)
-                  }
-                  formID={formID}
-                  formTitle={form.titleEn}
-                />
-              )}
-            </div>
-            {props.saveAndResume && dirty ? (
-              <SaveAndResume
-                formTitleEn={form.titleEn}
-                formTitleFr={form.titleFr}
-                securityAttribute={securityAttribute}
-                language={language as Language}
+            <FormActions
+              saveAndResumeEnabled={props.saveAndResume || false}
+              formId={formID}
+              language={language as Language}
+              form={form}
+              dirty={dirty}
+            >
+              <PrimaryFormButtons
+                isGroupsCheck={isGroupsCheck}
+                isShowReviewPage={isShowReviewPage}
+                groupsHeadingRef={groupsHeadingRef}
+                language={language}
                 formId={formID}
+                formTitle={form.titleEn}
+                getFormDelay={() =>
+                  isShowReviewPage
+                    ? getFormDelayWithGroups()
+                    : getFormDelayWithoutGroups(form.elements)
+                }
+                props={props}
               />
-            ) : null}
+            </FormActions>
           </form>
         </>
       }
     </>
   );
 };
-
-interface FormProps {
-  formRecord: PublicFormRecord;
-  initialValues?: Responses | undefined;
-  language: string;
-  isPreview?: boolean;
-  renderSubmit?: ({
-    validateForm,
-    fallBack,
-  }: {
-    validateForm: Validate["validateForm"];
-    fallBack?: () => JSX.Element;
-  }) => JSX.Element;
-  onSuccess: (id: string) => void;
-  children?: (JSX.Element | undefined)[] | null;
-  t: TFunction;
-  allowGrouping?: boolean | undefined;
-  groupHistory?: string[];
-  matchedIds?: string[];
-  saveSessionProgress: (language?: Language) => void;
-  saveAndResume?: boolean;
-}
 
 /**
  * This is the main Form component that wraps "InnerForm" withFormik hook, giving all of its components context
