+++ components/clientComponents/forms/SaveAndResume/SaveAndResume.tsx
@@ -2,7 +2,6 @@ import { useState } from "react";
 import { useTranslation } from "@i18n/client";
 
 import { type Language } from "@lib/types/form-builder-types";
-import { type SecurityAttribute } from "@lib/types";
 import { LinkButton } from "@serverComponents/globals/Buttons/LinkButton";
 import { SubmitButton as DownloadProgress } from "@clientComponents/globals/Buttons/SubmitButton";
 import { ConfirmDownload } from "./ConfirmDownload";
@@ -11,20 +10,18 @@ export const SaveAndResume = ({
   formId,
   formTitleEn,
   formTitleFr,
-  securityAttribute,
   language,
 }: {
   formId: string;
   formTitleEn: string;
   formTitleFr: string;
-  securityAttribute: SecurityAttribute;
   language: Language;
 }) => {
   const { t } = useTranslation(["review", "common"]);
   const [confirm, setConfirm] = useState(false);
 
   return (
-    <div className="sticky bottom-0 z-50 mt-10 flex border-t-2 border-gcds-blue-900 bg-gcds-blue-100 p-4">
+    <div>
       <DownloadProgress
         className="mr-4"
         type="button"
@@ -41,7 +38,6 @@ export const SaveAndResume = ({
         formId={formId}
         formTitleEn={formTitleEn}
         formTitleFr={formTitleFr}
-        securityAttribute={securityAttribute}
         language={language}
         open={confirm}
         handleClose={() => setConfirm(false)}
